IT-6541 | Add compact layouts to Tile component
In Tile, we want to have option to render it slightly different depending on the size of tile:
* Show only left icon
* Show only right icon
* Hide right icon - this was available through separate property "isRightIconHiddenOnMobile"
As container-type is applying some "different" styles to the element and changing it behaviour we need to only apply that when querying container size is really needed.
Also, we have been using "margins" for left and right icon which is unnecessary and instead we should use gap, which will make it easier to center elements for small variants.
DatePicker seems to be the only user of previous isRightIconHiddenOnMobile, so we should change the functionality to the `compactLayout` and use the `ICON_RIGHT_HIDDEN`